base/requirements.txt
Django>=3.0,<4.0
django-environ==0.9.0
psycopg2>=2.8
django-phonenumber-field==6.3.0
